requirements.txt
requests
python-dotenv
plotly.express
flask
pytest
gunicorn
numpy
pandas
cvxpy
seaborn
gspread
gspread_models
datetime

# Below data is currently defunct
# google.colab import auth
# google.auth import default
google.colab import userdata